string formatting assertion failure in MapEditContext
Fixed by tweaking format and explicitly casting.

If anyone actually *needs* to be able to select more than 2^31 - 1 map elements at once, I'm curious how they think anyone will be able to run their map :)

+++ src/MapEditor/MapEditContext.cpp
@@ -1265,7 +1265,7 @@ void MapEditContext::updateStatusText()
 
 	if (edit_mode_ != Mode::Visual && selection_.size() > 0)
 	{
-		mode += S_FMT(" (%lu selected)", selection_.size());
+		mode += S_FMT(" (%d selected)", (int) selection_.size());
 	}
 
 	MapEditor::setStatusText(mode, 1);
